feat: 重大重構 - 完整實現女性向AI聊天系統
🎯 核心系統實現:
- 完整的5級NSFW內容分級與智能路由 (OpenAI/Grok)
- 記憶管理系統：短期對話記憶 + 長期關係記憶
- 情感狀態追蹤：好感度系統 + 情緒狀態管理
- 女性向角色個性化：陸寒淵(霸道總裁) + 沈言墨(溫柔醫生)
- AI對話服務完整重構，支援真實API呼叫

📚 文檔架構重整:
- 更新SPEC.md：整合女性向系統完整設計，API優化到47個核心端點
- 新增MEMORY_GUIDE.md：記憶系統MVP實現指南
- 重寫NSFW_GUIDE.md：女性向NSFW功能實作指南
- 建立清晰文檔分工，消除重複維護負擔

🔧 技術架構優化:
- 統一日誌系統 (Logrus + 結構化輸出)
- 前端界面大幅強化 (對話/角色/設定頁面)
- CORS配置與環境變數完善
- API測試腳本增強，支援完整功能驗證

📊 變更規模: 30檔案 (+5,098/-797行)，新增4個核心服務模組

+++ handlers/emotion.go
@@ -189,7 +189,7 @@ func TriggerEmotionEvent(c *gin.Context) {
 
 	var req struct {
 		EventType string                 `json:"event_type" binding:"required"`
-		Intensity int                    `json:"intensity"`
+		Intensity float64                `json:"intensity"`
 		Context   map[string]interface{} `json:"context"`
 	}
